unet/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=0.13.0=pypi_0
albumentations=1.0.1=pyhd8ed1ab_0
appdirs=1.4.4=pyh9f0ad1d_0
argon2-cffi=20.1.0=pypi_0
async-generator=1.10=pypi_0
attrs=21.2.0=pypi_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
bleach=3.3.0=pypi_0
blosc=1.21.0=h0e60522_0
brotli=1.0.9=h8ffe710_5
brotli-bin=1.0.9=h8ffe710_5
brotlipy=0.7.0=py39hb82d6ee_1001
bzip2=1.0.8=h8ffe710_4
ca-certificates=2021.5.30=h5b45459_0
cachetools=4.2.2=pypi_0
certifi=2021.5.30=py39hcbf5309_0
cffi=1.14.5=py39h0878f49_0
cfitsio=3.470=h0af3d06_7
chardet=4.0.0=py39hcbf5309_1
charls=2.2.0=h39d44d4_0
cloudpickle=1.6.0=py_0
colorama=0.4.4=pyh9f0ad1d_0
conda=4.10.3=py39hcbf5309_0
conda-package-handling=1.7.3=py39hb3671d1_0
cryptography=3.4.7=py39hd8d06c1_0
cycler=0.10.0=py_2
cytoolz=0.11.0=py39hb82d6ee_3
dask-core=2021.6.2=pyhd8ed1ab_0
debugpy=1.3.0=py39h415ef7b_0
decorator=5.0.9=pyhd8ed1ab_0
defusedxml=0.7.1=pypi_0
entrypoints=0.3=pypi_0
freeglut=3.2.1=h0e60522_2
freetype=2.10.4=h546665d_1
fsspec=2021.6.1=pyhd8ed1ab_0
geos=3.9.1=h39d44d4_2
giflib=5.2.1=h8d14728_2
google-auth=1.32.1=pypi_0
google-auth-oauthlib=0.4.4=pypi_0
grpcio=1.38.1=pypi_0
icu=68.1=h0e60522_0
idna=2.10=pyh9f0ad1d_0
imagecodecs=2021.6.8=py39h166567b_0
imageio=2.9.0=py_0
imgaug=0.4.0=py_1
intel-openmp=2021.3.0=h57928b3_3372
ipykernel=6.0.1=py39h832f523_0
ipython=7.25.0=py39h832f523_1
ipython_genutils=0.2.0=py_1
ipywidgets=7.6.3=pypi_0
jasper=2.0.14=h77af90b_2
jbig=2.1=h8d14728_2003
jedi=0.18.0=py39hcbf5309_2
jinja2=3.0.1=pypi_0
joblib=1.0.1=pypi_0
jpeg=9d=h8ffe710_0
jsonschema=3.2.0=pypi_0
jupyter_client=6.1.12=pyhd8ed1ab_0
jupyter_core=4.7.1=py39hcbf5309_0
jupyterlab-pygments=0.1.2=pypi_0
jupyterlab-widgets=1.0.0=pypi_0
jxrlib=1.1=h8ffe710_2
kiwisolver=1.3.1=py39h2e07f2f_1
lcms2=2.12=h2a16943_0
lerc=2.2.1=h0e60522_0
libaec=1.0.5=h39d44d4_0
libblas=3.9.0=9_mkl
libbrotlicommon=1.0.9=h8ffe710_5
libbrotlidec=1.0.9=h8ffe710_5
libbrotlienc=1.0.9=h8ffe710_5
libcblas=3.9.0=9_mkl
libclang=11.1.0=default_h5c34c98_1
libdeflate=1.7=h8ffe710_5
liblapack=3.9.0=9_mkl
liblapacke=3.9.0=9_mkl
libopencv=4.5.2=py39h41a9e09_1
libpng=1.6.37=h1d00b33_2
libprotobuf=3.16.0=h7755175_0
libsodium=1.0.18=h8d14728_1
libtiff=4.3.0=h0c97f57_1
libwebp-base=1.2.0=h8ffe710_2
libzopfli=1.0.3=h0e60522_0
locket=0.2.0=py_2
lz4-c=1.9.3=h8ffe710_0
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markdown=3.3.4=pypi_0
markupsafe=2.0.1=pypi_0
matplotlib-base=3.4.2=py39h581301d_0
matplotlib-inline=0.1.2=pyhd8ed1ab_2
menuinst=1.4.16=py39hde42818_1
mistune=0.8.4=pypi_0
mkl=2021.2.0=hb70f87d_389
msys2-conda-epoch=20160418=1
nbclient=0.5.3=pypi_0
nbconvert=6.1.0=pypi_0
nbformat=5.1.3=pypi_0
nest-asyncio=1.5.1=pypi_0
networkx=2.3=py_0
notebook=6.4.0=pypi_0
numpy=1.21.0=py39h6635163_0
oauthlib=3.1.1=pypi_0
olefile=0.46=pyh9f0ad1d_1
opencv=4.5.2=py39hcbf5309_1
opencv-python=4.5.2.54=pypi_0
openjpeg=2.4.0=hb211442_1
openssl=1.1.1k=h8ffe710_0
packaging=21.0=pyhd8ed1ab_0
pandas=1.3.0=pypi_0
pandocfilters=1.4.3=pypi_0
parso=0.8.2=pyhd8ed1ab_0
partd=1.2.0=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=8.3.0=py39h916092e_0
pip=21.1.3=pyhd8ed1ab_0
pooch=1.4.0=pyhd8ed1ab_0
prometheus-client=0.11.0=pypi_0
prompt-toolkit=3.0.19=pyha770c72_0
protobuf=3.17.3=pypi_0
py-opencv=4.5.2=py39h832f523_1
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycosat=0.6.3=py39hb82d6ee_1006
pycparser=2.20=pyh9f0ad1d_2
pygments=2.9.0=pyhd8ed1ab_0
pyopenssl=20.0.1=pyhd8ed1ab_0
pyparsing=2.4.7=pyh9f0ad1d_0
pyrsistent=0.18.0=pypi_0
pysocks=1.7.1=py39hcbf5309_3
python=3.9.6=h7840368_0_cpython
python-dateutil=2.8.1=py_0
python_abi=3.9=2_cp39
pytz=2021.1=pypi_0
pywavelets=1.1.1=py39h5d4886f_3
pywin32=300=py39hb82d6ee_0
pywinpty=1.1.3=pypi_0
pyyaml=5.4.1=py39hb82d6ee_0
pyzmq=22.1.0=py39he46f08e_0
qt=5.12.9=h5909a2a_4
requests=2.25.1=pyhd3deb0d_0
requests-oauthlib=1.3.0=pypi_0
rsa=4.7.2=pypi_0
ruamel_yaml=0.15.80=py39hb82d6ee_1004
scikit-image=0.18.1=py39h2e25243_0
scikit-learn=0.24.2=pypi_0
scipy=1.7.0=py39hc0c34ad_0
send2trash=1.7.1=pypi_0
setuptools=49.6.0=py39hcbf5309_3
shapely=1.7.1=py39haadaec5_5
six=1.16.0=pyh6c4a22f_0
snappy=1.1.8=ha925a31_3
sqlite=3.36.0=h8ffe710_0
tbb=2021.3.0=h2d74725_0
tensorboard=2.5.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.0=pypi_0
terminado=0.10.1=pypi_0
testpath=0.5.0=pypi_0
threadpoolctl=2.1.0=pypi_0
tifffile=2021.7.2=pyhd8ed1ab_0
tk=8.6.10=h8ffe710_1
toolz=0.11.1=py_0
torch=1.9.0=pypi_0
tornado=6.1=py39hb82d6ee_1
tqdm=4.61.2=pyhd8ed1ab_0
traitlets=5.0.5=py_0
typing-extensions=3.10.0.0=pypi_0
tzdata=2021a=he74cb21_0
ucrt=10.0.20348.0=h57928b3_0
urllib3=1.26.6=pyhd8ed1ab_0
vc=14.2=hb210afc_5
vs2015_runtime=14.29.30037=h902a5da_5
wcwidth=0.2.5=pyh9f0ad1d_2
webencodings=0.5.1=pypi_0
werkzeug=2.0.1=pypi_0
wheel=0.36.2=pyhd3deb0d_0
widgetsnbextension=3.5.1=pypi_0
win_inet_pton=1.1.0=py39hcbf5309_2
wincertstore=0.2=py39hcbf5309_1006
xz=5.2.5=h62dcd97_1
yaml=0.2.5=he774522_0
zeromq=4.3.4=h0e60522_0
zfp=0.5.5=h0e60522_5
zlib=1.2.11=h62dcd97_1010
zstd=1.5.0=h6255e5f_0
